refactor instruction-related pod types to separate submodule
